app/requirements.txt
Flask==2.3.2
openai==1.68.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
bcrypt==4.0.1
redis==5.0.8
Flask-SocketIO==5.3.6
langdetect==1.0.9
APScheduler==3.10.4
werkzeug==2.3.7